backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
gunicorn==21.2.0

# Authentication and security
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
google-auth==2.23.4

# Data validation and settings
pydantic
pydantic[email]
pydantic-settings

# Google Generative AI (Gemini)
google-generativeai==0.3.2

# Image processing for AI OCR
Pillow==10.1.0

# Environment configuration
python-dotenv==1.0.0

# Development tools (optional - for testing and development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Form data handling
python-multipart==0.0.20

